`Cell::partition_id` is stored as `int`
This is the MPI rank, which is stored as `int`.

Refs #90

+++ framework/graphs/kba_graph_partitioner.cc
@@ -83,7 +83,7 @@ KBAGraphPartitioner::KBAGraphPartitioner(const InputParameters& params)
   } // for each coordinate
 }
 
-std::vector<int64_t>
+std::vector<int>
 KBAGraphPartitioner::Partition(const std::vector<std::vector<uint64_t>>& graph,
                                const std::vector<Vector3>& centroids,
                                int number_of_parts)
@@ -93,7 +93,7 @@ KBAGraphPartitioner::Partition(const std::vector<std::vector<uint64_t>>& graph,
   OpenSnLogicalErrorIf(centroids.size() != graph.size(),
                        "Graph number of entries not equal to centroids' number of entries.");
   const size_t num_cells = graph.size();
-  std::vector<int64_t> pids(num_cells, 0);
+  std::vector<size_t> pids(num_cells, 0);
   for (size_t c = 0; c < num_cells; ++c)
   {
     const auto& point = centroids[c];
@@ -132,13 +132,13 @@ KBAGraphPartitioner::Partition(const std::vector<std::vector<uint64_t>>& graph,
 
   const auto pid_subsets = MakeSubSets(nx_ * ny_ * nz_, number_of_parts);
 
-  std::vector<int64_t> real_pids(num_cells, 0);
+  std::vector<int> real_pids(num_cells, 0);
   for (size_t c = 0; c < num_cells; ++c)
   {
     for (int p = 0; p < number_of_parts; ++p)
     {
       if (pids[c] >= pid_subsets[p].ss_begin and pids[c] <= pid_subsets[p].ss_end)
-        real_pids[c] = static_cast<int64_t>(p);
+        real_pids[c] = p;
     }
   }
